+++ src/main/kotlin/com/github/dinbtechit/vscodetheme/actions/DonateAction.kt
@@ -1,0 +1,14 @@
+package com.github.dinbtechit.vscodetheme.actions
+
+import com.github.dinbtechit.vscodetheme.icons.VSCodeIcons
+import com.intellij.ide.BrowserUtil
+import com.intellij.openapi.actionSystem.AnActionEvent
+import com.intellij.openapi.project.DumbAwareAction
+
+class DonateAction: DumbAwareAction("Donate", "", VSCodeIcons.BuyMeCoffee) {
+
+    override fun actionPerformed(e: AnActionEvent) {
+        BrowserUtil.open("https://www.buymeacoffee.com/dinbtechit")
+    }
+
+}