+++ src/main/kotlin/com/github/dinbtechit/vscodetheme/icons/VSCodeIcons.kt
@@ -4,4 +4,5 @@ import com.intellij.openapi.util.IconLoader
 
 object VSCodeIcons {
     val Logo = IconLoader.getIcon("icons/logo-16.svg", javaClass)
+    val BuyMeCoffee = IconLoader.getIcon("icons/buy-me-coffee.svg", javaClass)
 }